+++ app/controllers/trackables_controller.ts
@@ -10,7 +10,6 @@ export default class TrackablesController {
   public async create_trackable({ request, response, auth }: HttpContext) {
     const user = await auth.authenticate()
     const data = request.only([
-      'user_id',
       'photo',
       'category',
       'type',
@@ -47,6 +46,7 @@ export default class TrackablesController {
         ...data,
         index: nbrTrackable + 1,
         id,
+        user_id: user.id,
         photos: JSON.stringify(photosUrl),
       })
       return response.status(201).json({
@@ -65,35 +65,34 @@ export default class TrackablesController {
     const user = await auth.authenticate()
 
     try {
-      const { page, limit, order_by, categories, types, name, trackable_id } = request.qs()
+      const {
+        page = 1,
+        limit = 10,
+        order_by = 'date_asc',
+        categories,
+        types,
+        name,
+        trackable_id,
+      } = request.qs()
       const _categories = JSON.parse(categories)
       const _type = JSON.parse(types)
-      if (trackable_id) {
-        const trackable = await Trackable.find(trackable_id)
-
-        if (!trackable) {
-          return response.status(404).json({
-            message: 'Trackable non trouvé',
-          })
-        }
-
-        return response.status(200).json({
-          message: 'Trackable récupéré avec succès',
-          data: trackable,
-        })
-      }
 
       let query = db
         .from(Trackable.table)
         .select('*')
         .innerJoin('users', 'users.id', 'trackables.user_id')
         .where('user_id', user.id)
+
+      if (trackable_id) {
+        query = query.where('id', trackable_id)
+      }
       if (_categories.length) {
         query = query.whereIn('category', _categories)
       }
       if (order_by) {
         query = applyOrderBy(query, order_by, Trackable.table)
       }
+
       if (_type.length) {
         query = query.whereIn('type', _type)
       }
@@ -123,8 +122,9 @@ export default class TrackablesController {
     }
   }
 
-  public async update_trackable({ request, response }: HttpContext) {
+  public async update_trackable({ request, response, auth }: HttpContext) {
     const trackable_id = request.input('trackable_id')
+    const user = await auth.authenticate()
     const body = request.all()
     try {
       const trackable = await Trackable.find(trackable_id)
@@ -134,7 +134,6 @@ export default class TrackablesController {
         })
       }
       const data = request.only([
-        'user_id',
         'category',
         'type',
         'name',
@@ -149,6 +148,11 @@ export default class TrackablesController {
         'qr_code',
         'index',
       ])
+      if (user.id !== trackable.user_id) {
+        return response.status(403).json({
+          message: "Vous n'avez pas le droit d'editer ce trackable",
+        })
+      }
       let urls: string[]
       urls = await updateFiles({
         request,
@@ -181,7 +185,8 @@ export default class TrackablesController {
     }
   }
 
-  public async delete_trackable({ params, response }: HttpContext) {
+  public async delete_trackable({ params, response, auth }: HttpContext) {
+    const user = await auth.authenticate()
     try {
       const trackable = await Trackable.find(params.trackable_id)
 
@@ -191,8 +196,12 @@ export default class TrackablesController {
         })
       }
 
+      if (user.id !== trackable.user_id) {
+        return response.status(403).json({
+          message: "Vous n'avez pas le droit de supprimer ce trackable",
+        })
+      }
       await trackable.delete()
-
       return response.status(200).json({
         message: 'Trackable supprimé avec succès',
       })
